Rename CloudFerro sync script
