+++ deafrica/data/cdse_pipelines/sentinel_sync_cloudferro.py
