+++ setup.cfg
@@ -73,8 +73,7 @@ console_scripts =
     s1-gap-filler = deafrica.monitoring.s1_gap_filler:cli
     s1-cdse-pipeline = deafrica.data.cdse_pipelines.s1:cli
     s1-generate-xml-metadata = deafrica.data.cdse_pipelines.s1_generate_xml_metadata:cli
-    s1-sync-cloudferro = deafrica.data.cdse_pipelines.s1_sync_cloudferro:cli
-    sentinel-sync-cloudferro = deafrica.data.cdse_pipelines.s1_sync_cloudferro:cli
+    sentinel-sync-cloudferro = deafrica.data.cdse_pipelines.sentinel_sync_cloudferro:cli
     landsat-gap-report = deafrica.monitoring.landsat_gap_report:cli
     landsat-gap-filler = deafrica.monitoring.landsat_gap_filler:cli
     s2-gap-report = deafrica.monitoring.s2_gap_report:cli
